Extract legend composable from DayCalendarScreen and put it on composables package
Change the legned in CalendarScreen for the exact same of DayCalendarScreen
