+++ app/src/main/java/com/android/sample/ui/calendar/CalendarScreen.kt
@@ -29,6 +29,8 @@ import com.android.sample.model.calendar.CalendarViewModel
 import com.android.sample.model.workout.Workout
 import com.android.sample.model.workout.WorkoutType
 import com.android.sample.model.workout.WorkoutViewModel
+import com.android.sample.ui.composables.BottomBar
+import com.android.sample.ui.composables.Legend
 import com.android.sample.ui.composables.TopBar
 import com.android.sample.ui.navigation.NavigationActions
 import com.android.sample.ui.navigation.Screen
@@ -149,11 +151,15 @@ fun CalendarScreen(
 
   val workoutsByDate = workouts.groupBy { it.workout.date.toLocalDate().toKotlinLocalDate() }
 
-  Scaffold(topBar = { TopBar(navigationActions, R.string.calendar_title) }) { innerPadding ->
+  Scaffold(topBar = { TopBar(navigationActions, R.string.calendar_title) },
+      bottomBar = { BottomBar(navigationActions) }) { innerPadding ->
     Column(modifier = Modifier.fillMaxSize().padding(innerPadding)) {
-      Spacer(modifier = Modifier.height(8.dp))
+      Legend()
 
-      Legend(modifier = Modifier.testTag("legend"))
+        Divider(
+            color = Color.LightGray,
+            thickness = 1.dp,
+            modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp))
 
       LazyColumn(
           state = lazyListState,
@@ -179,31 +185,6 @@ private fun generateDateRange(
     count: Int
 ): List<kotlinx.datetime.LocalDate> {
   return List(count) { daysToAdd -> startDate.plus(daysToAdd.toLong(), DateTimeUnit.DAY) }
-}
-
-@Composable
-fun Legend(modifier: Modifier = Modifier) {
-  Row(
-      modifier = modifier.fillMaxWidth().padding(horizontal = 16.dp, vertical = 2.dp),
-      horizontalArrangement = Arrangement.SpaceEvenly,
-      verticalAlignment = Alignment.CenterVertically) {
-        LegendItem(
-            color = PastelRed,
-            label = "Bodyweight",
-            modifier = Modifier.testTag("legendBodyweight"))
-        LegendItem(color = PastelBlue, label = "Yoga", modifier = Modifier.testTag("legendYoga"))
-      }
-}
-
-@Composable
-fun LegendItem(color: Color, label: String, modifier: Modifier = Modifier) {
-  Row(
-      modifier = modifier,
-      verticalAlignment = Alignment.CenterVertically,
-      horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-        Box(modifier = Modifier.size(16.dp).background(color, shape = MaterialTheme.shapes.small))
-        Text(text = label, style = MaterialTheme.typography.bodySmall)
-      }
 }
 
 @Composable
